Add option to ignore whether client is ingame
Add an option to explicitly ignore whether IsClientInGame is true.
This can be useful eg. when using IsValidClient inside other general use
functions such as admin checks that might get called by
OnClientAuthorized (which can be called before IsClientInGame is true)
and the client index is already guaranteed to be valid.