Add NEO_MAXPLAYERS constant
SourceMod's default MAXPLAYERS constant of 64 is double the size of NT's MAXPLAYERS (32), so most plugin roll their own constant to avoid doubly allocations etc. Adding this to the neotokyo.inc itself to avoid this code repetition. Nearly all plugins defined this as NEO_MAX_PLAYERS or similar instead of NEO_MAXPLAYERS, so any breakage from a redefined constant should be minimal (as of writing this, no NT plugins hosted on GitHub clash with this definition name).